+++ app/src/main/java/org/oppia/app/player/state/itemviewmodel/ContentViewModel.kt
@@ -1,4 +1,4 @@
-package org.oppia.app.player.state
+package org.oppia.app.player.state.itemviewmodel
 
 import androidx.lifecycle.ViewModel
 import org.oppia.app.fragment.FragmentScope
@@ -6,6 +6,8 @@ import javax.inject.Inject
 
 /** [ViewModel] for content-card state. */
 @FragmentScope
-class NumericInputInteractionViewModel @Inject constructor() : ViewModel() {
-  var placeholder = ""
+class ContentViewModel @Inject constructor() : ViewModel() {
+
+  var contentId = ""
+  var htmlContent = ""
 }
