+++ app/src/main/java/org/oppia/app/player/state/itemviewmodel/NumericInputInteractionViewModel.kt
@@ -1,4 +1,4 @@
-package org.oppia.app.player.state
+package org.oppia.app.player.state.itemviewmodel
 
 import androidx.lifecycle.ViewModel
 import org.oppia.app.fragment.FragmentScope
@@ -6,8 +6,6 @@ import javax.inject.Inject
 
 /** [ViewModel] for content-card state. */
 @FragmentScope
-class ContentViewModel @Inject constructor() : ViewModel() {
-
-  var contentId = ""
-  var htmlContent = ""
+class NumericInputInteractionViewModel @Inject constructor() : ViewModel() {
+  var placeholder = ""
 }
