+++ app/src/main/java/org/oppia/app/player/state/listener/InteractionAnswerRetiever.kt
@@ -1,4 +1,4 @@
-package org.oppia.app.player.state
+package org.oppia.app.player.state.listener
 
 import org.oppia.app.model.InteractionObject
