+++ app/src/main/res/layout/numeric_input_interaction_item.xml
@@ -14,7 +14,7 @@
     android:layout_marginEnd="4dp"
     android:layout_marginBottom="4dp"
     android:orientation="vertical">
-    <org.oppia.app.player.state.NumericInputInteractionView
+    <org.oppia.app.player.state.customview.NumericInputInteractionView
       android:id="@+id/numeric_input_interaction_view"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
